+++ PureInvoke/Functions/Get-Kernel32.ps1
@@ -1,0 +1,27 @@
+
+function Get-Kernel32
+{
+    Set-StrictMode -Version 'Latest'
+    Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
+
+    $kernel32 = @"
+        [DllImport("kernel32.dll", SetLastError=true)]
+        public static extern bool FindClose(IntPtr hFindFile);
+
+        [DllImport("kernel32.dll", CharSet=CharSet.Unicode, SetLastError=true)]
+        public static extern IntPtr FindFirstFileNameW(string lpFileName, uint dwFlags, ref uint stringLength,
+            StringBuilder linkName);
+
+        [DllImport("kernel32.dll", CharSet=CharSet.Unicode, SetLastError=true)]
+        public static extern bool FindNextFileNameW(IntPtr hFindStream, ref uint stringLength, StringBuilder linkName);
+
+        [DllImport("kernel32.dll", CharSet=CharSet.Unicode, SetLastError=true)]
+        public static extern bool GetVolumePathName(string lpszFileName, [Out] StringBuilder lpszVolumePathName,
+            uint cchBufferLength);
+
+        [DllImport("kernel32.dll")]
+        public static extern IntPtr LocalFree(IntPtr hMem);
+"@
+
+    return Add-PInvokeType -DllName 'Kernel32' -Definition $kernel32
+}