+++ src/WinNT/LUID.cs
@@ -1,9 +1,0 @@
-
-namespace PureInvoke.WinNT
-{
-	public struct LUID
-	{
-		public uint LowPart;
-		public int HighPart;
-	}
-}